\name{check_replicate_paperword}
\alias{check_replicate_paperword}
\title{Check repeat english word(s) in text string}
\usage{
check_replicate_paperword(txt) 
}
\description{
Check repeat english word(s) in text string. This function be finished in 2013. 
}
\arguments{
	\item{txt} {a string vector.}
}


\examples{
txt <- "This is is a place. This is is a place"
check_replicate_paperword(txt)

}
