\name{day.value.ibutton}
\alias{day.value.ibutton}
\title{Calculate the day-mean temperature or humidity based data from ibutton}
\usage{
day.value.ibutton(filename = filename,humidity.correct=FALSE)
}
\description{
This function can be used to calculate the day-mean temperature or humidity based data from ibutton equipment.
}
\arguments{
	\item{filename} {the file name from ibutton equipment.}
	\item{humidity.correct} { relative humidity correction when the strange value is over 100.}
}
