
\name{draw_colors657}
\alias{draw_colors657}
\title{The function displays and labels all 657 built-in R colors in a grid layout.}
\usage{
draw_colors657(...) 
}
\description{
The function displays and labels all 657 built-in R colors in a grid layout
}
\arguments{
	\item{...} {further argument pass to plot function.}
 
}


\examples{
draw_colors657() 

}
