\name{draw.rletter}
\alias{draw.rletter}
\title{Draw a R letter using R}
\usage{
draw.rletter(col='blue',...)
}
\description{
Draw a R letter using R.
}
\arguments{
	\item{col} {Default is blue colour.}
	\item{...} {further arguments pass to the \code{\link{polygon}} function. }
}


\examples{
draw.rletter() 

}
