\name{escience.review}
\alias{escience.review}
\title{Fetch the number of visitors}
\usage{
escience.review(author)
}
\description{
This function can be used to extract fetch the number of visitors on escience website (http://www.escience.cn/people/xxx/index.html).
}
\arguments{
	\item{author} {the author name. }
}

\examples{

escience.review(author='liuguofang')

}
