\name{installed.pack.search}
\alias{installed.pack.search}
\title{Search installed package based on a string}
\usage{
installed.pack.search(string)
}
\description{
This function can be used to search installed package based on a string.
}

\arguments{
	\item{string} {a string.}
}

\examples{
installed.pack.search('mess')

installed.pack.search()
}
