\name{north.direction}
\alias{north.direction}
\title{Create a legend with north direction}
\usage{
north.direction(x.change = 0, y.change = 0)
}
\description{
Create a legend with north direction.
}
\arguments{
	\item{x.change} {a x change from topright to other region of figure. The value ranges 0-1.}
	\item{x.change} {a y change from topright to other region of figure. The value ranges 0-1.}
}


\examples{
plot.new()
north.direction()

}
