\name{province_area}
\alias{province_area}
\title{Calculate the province or city area}
\usage{
province.area(province)
}
\description{
Calculate the province or city area. The province.area function was written in 2013. 
}
\arguments{
	\item{province} {a province name.}
}

\examples{

 province_area('新疆')

 province.area('河南省')

}
