\name{twopoints.dist}
\alias{twopoints.dist}
\title{Calculate the distance between two points on earth}
\usage{
twopoints.dist(lat1, lon1, lat2, lon2) 
}
\description{
Calculate the distance between two points on earth.
}
\arguments{
	\item{lat1} {a latitude value of a point.}
	\item{lon1} {a longitude value of a point.}
  \item{lat2} {a latitude value of another point.}
  \item{lon2} {a latitude value of another point.}
  
}


\examples{
twopoints.dist(39,110,40,110)

}
