\name{temp.dayscale.plot.ibutton}
\alias{temp.dayscale.plot.ibutton}
\title{Plot dayscale temperature and humidity}
\usage{
temp.dayscale.plot.ibutton(temp=temperature,humidity=NULL,start.time='07:00')
}
\description{
This function can be used to plot dayscale temperature and humidity data from ibutton equipment.  
}
\arguments{
	\item{temp} {the name of temperature file.}
	\item{humidity} {the name of humidity file. Default is drawing dayscale temperature curve. }
  	\item{res}{start.time } {the start point of temperature or humidity graph}
        
}
